requirements.txt
backports-abc==0.4
beautifulsoup4==4.3.2
boto==2.38.0
Bottleneck==1.0.0
bz2file==0.98
colorama==0.3.3
contextlib2==0.4.0
cycler==0.10.0
Cython==0.23.4
decorator==4.0.9
funcy==1.6
future==0.15.2
gensim==0.12.3
ipykernel==4.3.1
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
joblib==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.1.1
jupyter-core==4.1.0
liac-arff==2.0.2
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.1
matplotlib-tests==1.5.1
memory-profiler==0.32
mistune==0.7.1
mock==1.3.0
natsort==4.0.3
nbconvert==4.1.0
nbformat==4.0.1
nltk==3.2
nose==1.3.7
notebook==4.1.0
numexpr==2.4.4
numpy==1.10.4
pandas==0.17.1
path.py==8.1.2
pbr==1.8.1
pickleshare==0.6
Pillow==3.0.0
psutil==2.2.1
py==1.4.30
py-common-subseq==0.1.4
pycairo==1.10.0
Pygments==2.0.2
pyLDAvis==1.3.0
pyparsing==2.1.0
PySide==1.2.2
pytest==2.8.2
python-dateutil==2.5.0
pytz==2015.7
PyYAML==3.11
pyzmq==15.0.0
qtconsole==4.2.0
readability==0.1
scikit-learn==0.17.1
scipy==0.16.1
seaborn==0.6.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.2.1
SQLAlchemy==1.0.9
statsmodels==0.6.1
tables==3.2.2
textblob==0.9.0
tornado==4.3
traitlets==4.1.0
vaderSentiment==0.5
